remove dart:html usage and used package:web instead
