+++ dart/lib/src/event_processor/enricher/enricher_event_processor.dart
@@ -1,7 +1,6 @@
 import '../../event_processor.dart';
 import '../../sentry_options.dart';
 import 'io_enricher_event_processor.dart'
-    if (dart.library.html) 'html_enricher_event_processor.dart'
     if (dart.library.js_interop) 'web_enricher_event_processor.dart';
 
 abstract class EnricherEventProcessor implements EventProcessor {
