+++ dart/lib/src/origin.dart
@@ -1,3 +1,2 @@
 export 'origin_io.dart'
-    if (dart.library.html) 'origin_html.dart'
     if (dart.library.js_interop) 'origin_web.dart';
