--- dart/lib/src/origin_html.dart
@@ -1,4 +1,0 @@
-import 'dart:html';
-
-/// request origin, used for browser stacktrace
-String get eventOrigin => '${window.location.origin}/';
