+++ dart/lib/src/origin_web.dart
@@ -1,5 +1,3 @@
-// We would lose compatibility with old dart versions by adding web to pubspec.
-// ignore: depend_on_referenced_packages
 import 'package:web/web.dart';
 
 /// request origin, used for browser stacktrace
