+++ dart/lib/src/platform/platform.dart
@@ -1,8 +1,7 @@
 import 'dart:typed_data';
 
-import '_io_platform.dart'
-    if (dart.library.html) '_html_platform.dart'
-    if (dart.library.js_interop) '_web_platform.dart' as platform;
+import '_io_platform.dart' if (dart.library.js_interop) '_web_platform.dart'
+    as platform;
 
 const Platform instance = platform.instance;
