+++ dart/pubspec.yaml
@@ -25,6 +25,7 @@ dependencies:
   stack_trace: ^1.10.0
   uuid: '>=3.0.0 <5.0.0'
   collection: ^1.16.0
+  web: ^1.1.0
 
 dev_dependencies:
   build_runner: ^2.3.0
