+++ dart/test/event_processor/enricher/web_enricher_test.dart
@@ -2,9 +2,7 @@
 library;
 
 import 'package:sentry/sentry.dart';
-import 'package:sentry/src/event_processor/enricher/html_enricher_event_processor.dart'
-    if (dart.library.html) 'package:sentry/src/event_processor/enricher/html_enricher_event_processor.dart'
-    if (dart.library.js_interop) 'package:sentry/src/event_processor/enricher/web_enricher_event_processor.dart';
+import 'package:sentry/src/event_processor/enricher/web_enricher_event_processor.dart';
 import 'package:test/test.dart';
 
 import '../../mocks.dart';
