+++ file/lib/src/sentry_file_extension.dart
@@ -1,8 +1,6 @@
 // ignore_for_file: invalid_use_of_internal_member
 
-import 'dart:io'
-    if (dart.library.html) 'dart:html'
-    if (dart.library.js_interop) 'dart:js_interop';
+import 'dart:io' if (dart.library.js_interop) 'dart:js_interop';
 
 import 'package:meta/meta.dart';
 import 'package:sentry/sentry.dart';
@@ -29,9 +27,7 @@ extension SentryFileExtension on File {
   ///
   /// All the copy, create, delete, open, rename, read, and write operations are
   /// supported.
-  File sentryTrace({
-    @internal Hub? hub,
-  }) {
+  File sentryTrace({@internal Hub? hub}) {
     final _hub = hub ?? HubAdapter();
 
     if (_hub.options.platformChecker.isWeb ||
