+++ flutter/lib/src/event_processor/url_filter/url_filter_event_processor.dart
@@ -1,6 +1,5 @@
 import '../../../sentry_flutter.dart';
 import 'io_url_filter_event_processor.dart'
-    if (dart.library.html) 'html_url_filter_event_processor.dart'
     if (dart.library.js_interop) 'web_url_filter_event_processor.dart';
 
 abstract class UrlFilterEventProcessor implements EventProcessor {
