+++ flutter/lib/src/integrations/connectivity/web_connectivity_provider.dart
@@ -1,7 +1,5 @@
 import 'dart:async';
 
-// We would lose compatibility with old dart versions by adding web to pubspec.
-// ignore: depend_on_referenced_packages
 import 'package:web/web.dart' as web;
 
 import 'connectivity_provider.dart';
