+++ flutter/lib/src/web/script_loader/script_dom_api.dart
@@ -1,7 +1,6 @@
 import 'package:meta/meta.dart';
 
 export 'noop_script_dom_api.dart'
-    if (dart.library.html) 'html_script_dom_api.dart'
     if (dart.library.js_interop) 'web_script_dom_api.dart';
 
 @internal
