+++ flutter/pubspec.yaml
@@ -29,6 +29,7 @@ dependencies:
   ffi: ^2.0.0
   file: '>=6.1.4'
   collection: ^1.16.0
+  web: ^1.1.0
 
 dev_dependencies:
   build_runner: ^2.4.2
